% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_bb.R
\name{city_bb}
\alias{city_bb}
\title{city_bb}
\usage{
city_bb(city, expand = 0)
}
\arguments{
\item{city}{(in standard two-letter \code{bikedat} form: "lo", "ny", "bo",
"ch", "dc", or "la").}

\item{expand}{Relative factor by which to expand the bbox}
}
\value{
Numeric vector of four values defining the bounding box
}
\description{
Return the bounding box of bicycle stations from one city
}
