% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{get_all_networks}
\alias{get_all_networks}
\title{get_all_networks}
\usage{
get_all_networks()
}
\value{
Nothing (data are written to disk).
}
\description{
Extract street networks for all \code{bikedata} cities, and save each one in
a \code{streetnet_XX.rda} file.
}
