% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{get_streetnet}
\alias{get_streetnet}
\title{get_streetnet}
\usage{
get_streetnet(city)
}
\arguments{
\item{city}{(in standard two-letter \code{bikedat} form: "lo", "ny", "bo",
"ch", "dc", or "la").}
}
\description{
Get the street network for a particular city
}
