% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet-package.R
\docType{package}
\name{streetnet}
\alias{streetnet}
\alias{streetnet-package}
\title{Extract OSM street networks for bikedata cities}
\description{
The \code{bikedata} cities currently include
\itemize{
\item Santander Cycles London, U.K.
\item citibike New York City NY, U.S.A.
\item Divvy Chicago IL, U.S.A.
\item Capital BikeShare Washingon DC, U.S.A.
\item Hubway Boston MA, U.S.A.
\item Metro Los Angeles CA, U.S.A.
\item Indego Philadephia PA, U.S.A.
}
}
\author{
Mark Padgham
}
